Fix the alien spawning logic.
The problem was actually the way I expired the aliens: trying to be clever and copying the last array element over the one we were expiring, when I should have just been deleting the original array element.

Modified the code to delete the array element for the alien we're expiring & it fixes the problem.